Changement d'API : lufthansa -> flightAPI
Changement d'API et suppression des fichiers inutiles
