+++ .gitignore
@@ -1,2 +1,3 @@
 
 __pycache__/API.cpython-310.pyc
+*.pyc
